// src/app/services/models/turnover-report-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DepartmentTurnoverDto } from '../models/department-turnover-dto';
import { MonthlyTurnoverDto } from '../models/monthly-turnover-dto';
export interface TurnoverReportDto {
  departmentTurnover?: Array<DepartmentTurnoverDto>;
  departures?: number;
  endEmployeeCount?: number;
  monthlyTurnover?: Array<MonthlyTurnoverDto>;
  newHires?: number;
  startEmployeeCount?: number;
  turnoverRate?: number;
  year?: number;
}